fix: policy Order for admin & buyer

+++ resources/views/livewire/orders/delete.blade.php
@@ -13,6 +13,8 @@ new class extends Component {
 
     public function deleteOrder()
     {
+        $this->authorize('delete', $this->order);
+        
         $this->order->delete();
 
         Flux::toast(variant: 'danger', heading: 'Pedido eliminado', text: 'El pedido fue eliminado exitosamente');
